Reenable dataflow tests
The tests still appear to need a good amount of cleanup, and much of the testing is done in outerloop tests that still don't run.  This just re-enables the existing innerloop tests so that at least some tests run as part of builds / CI.
